Initial commit for base config files.
I've added on file to hold all default configurations that are being repeated for municipalities. There is now an exported function that will return the config needed for a specified row. This function also accepts override options for the configs for a given row. This will let us not worry about the configs until we really need to edit values for a given extraction.

I've added two interfaces that represent rows: BaseRow and Row. Row extends BaseRow. BaseRow represents the common fields that bow regular rows and rows for `Titulares` contain.