Add some quality of life improvements
The process for running this script and creating a specific config per municipality was beginning to become troublesome and prone to errors. I've made a couple of changes to help mitigate this.

I'm now importing the getConfig function that will use a baseConfig object. Passing the precint and the name of the municipality will now retun a configuration that is made for the municipality. The first argument still expects the relative path the the image and is constructed by using the new municipio and precinto variables. These two variables are used to build the location of the image to process and the folder that is created to place all the assests into.

I've added a number of console outputs to see where in the process the script is and to ensure that there isn't some stopping the script. For example waiting for the OCR service.

I also changed some of the paths and the way we construct our output. We are uploading the script output into Google Drive with a specific directory structure. The output now creates this structure so that the contributor can just drag and drop it into the shared Drive folder.

At the end of the run we are now also copying the original ballot image that we were processing into the output folder for comparison and safe keeping.