Initial commit of a for a base config
This is the second try at a base config file and it's primarily for some quality of life for contributors.

There are a number of things in this file:
- A new Enum that should be used to identify which configuration you want to use, Legislative or Municipal.
- A getConfig factory like function. This function will accept the municipality and precinct that the user is working on as well as the type of config that is needed. The municipio and precinto are used to also create the paths for the original images to process as well as the output folder to use for the processed assets. This function also default to use the Legislative config.

It's important to note that these configs are just a base line for conveniece. Any tweeks to the values should be does in the tools main entrypoint.