Refactor the use of base configs
I'm changing how we use the getConfig function to reflect changes in its signature. We now are importing the ConfigType enum and passing it into the getConfig function. This will let us use both base configuration for municipal and ligislative assets.

We are also not constructing the paths for the orignial image or the output directory since this is now done in the getConfig funtion.